TLDR; Crappy workaround for slider not working on certain phones.
Used https://hover-pointer-media-query.glitch.me/ to determine
that some phones don't report their hover/pointer capabilities
properly.
Had to resort to setting overflow-x: scroll constantly but that caused
ugly scrollbars to show on desktop.
In the end, I just set overflow-x: scroll when the screen width is that
of a mobile device. Not the best.